dev merger

+++ src/reflex.API/reflex.Domain/Entities/BaseEntities/BaseEntity.cs
@@ -4,5 +4,5 @@ namespace reflex.Domain;
 
 public class BaseEntity
 {
-    public BigInteger id { get; set;}
+    public int id { get; set;}
 }
